+++ app/src/main/java/de/christian2003/chaching/plugin/presentation/view/type/TypeScreen.kt
@@ -42,6 +42,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.dimensionResource
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.unit.dp
 import de.christian2003.chaching.R
 import de.christian2003.chaching.domain.type.TypeIcon
@@ -68,9 +69,9 @@ fun TypeScreen(
             TopAppBar(
                 title = {
                     Text(
-                        viewModel.name.ifEmpty {
-                            viewModel.namePlaceholder
-                        }
+                        text = viewModel.name.ifEmpty { viewModel.namePlaceholder },
+                        maxLines = 1,
+                        overflow = TextOverflow.Ellipsis
                     )
                 },
                 navigationIcon = {